Final commit with relationship

+++ app/Http/Controllers/ListingController.php
@@ -47,6 +47,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         // create the listing in the database
         Listing::create($formFields);
 
@@ -60,6 +62,11 @@ class ListingController extends Controller
 
     // Update Listing Data
     public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         // validation for form
         $formFields = $request->validate([
             'title' => 'required',
@@ -84,8 +91,18 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    // Manage listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
